requirements.txt
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
pyserial==3.5
winaudio==1.0.6
pyqt-tools
pyqt5designer
PyQt5-stubs==5.15.2.0
